Added ComparisonQuizCore skipCost to remote config api

+++ comparison-quiz/src/main/java/com/infinitepower/newquiz/comparison_quiz/ui/ComparisonQuizViewModel.kt
@@ -149,7 +149,7 @@ class ComparisonQuizViewModel @Inject constructor(
 
         _uiState.update { currentState ->
             currentState.copy(
-                userDiamonds = user.data.diamonds.toInt(),
+                userDiamonds = user.data.diamonds,
                 userDiamondsLoading = false
             )
         }
